refactor: gallery folder structure for better organization
