--- src/components/Gallery/Gallery.module.scss
@@ -1,4 +1,0 @@
-.heading {
-  text-align: center;
-  margin-bottom: 0.5em;
-}
