+++ src/components/UI/Error/Error.jsx
@@ -1,5 +1,5 @@
-import style from './Error.module.scss'
 import PropTypes from 'prop-types'
+import style from './Error.module.scss'
 
 const Error = ({ children }) => {
   return <p className={style.error}>{children}</p>
