+++ src/components/UI/Grid/Grid.jsx
@@ -1,5 +1,5 @@
-import style from './Grid.module.scss'
 import PropTypes from 'prop-types'
+import style from './Grid.module.scss'
 
 const Grid = ({ children }) => {
   return <div className={style.grid}>{children}</div>
