+++ src/components/gallery/PhotoCard/PhotoCard.module.scss
@@ -2,7 +2,7 @@
 @import '/src/styles/variables';
 @import '/src/styles/media-queries';
 
-.card {
+.photo {
   width: 100%;
   height: 250px;
   border-radius: 5px;
