+++ src/pages/HomePage.jsx
@@ -1,9 +1,10 @@
-import Gallery from '../components/Gallery/Gallery'
+import Gallery from '../components/gallery/Gallery'
 import Container from '../components/UI/Container/Container'
 
 const HomePage = () => {
   return (
     <Container>
+      <h1>Photo Gallery</h1>
       <Gallery />
     </Container>
   )
