+++ src/styles/global.scss
@@ -1,7 +1,9 @@
 @import './variables';
 @import './media-queries';
 
-*, *::before, *::after {
+*,
+*::before,
+*::after {
   box-sizing: border-box;
 }
 
@@ -44,6 +46,11 @@ h5,
 h6 {
   margin-top: 0;
   margin-bottom: 0;
+}
+
+h1 {
+  text-align: center;
+  margin-bottom: 0.5em;
 }
 
 h2 {
